move FallbackCounters to parser root dir
